requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile requirements.in
#
colorama==0.4.6
    # via tqdm
contourpy==1.3.3
    # via matplotlib
cycler==0.12.1
    # via matplotlib
filelock==3.20.0
    # via torch
fonttools==4.60.1
    # via matplotlib
fsspec==2025.10.0
    # via torch
imageio==2.37.0
    # via scikit-image
jinja2==3.1.6
    # via torch
kiwisolver==1.4.9
    # via matplotlib
lazy-loader==0.4
    # via scikit-image
markupsafe==3.0.3
    # via jinja2
matplotlib==3.10.7
    # via -r requirements.in
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via
    #   scikit-image
    #   torch
numpy==2.2.6
    # via
    #   -r requirements.in
    #   contourpy
    #   imageio
    #   matplotlib
    #   opencv-python
    #   scikit-image
    #   scipy
    #   tifffile
    #   torchvision
opencv-python==4.12.0.88
    # via -r requirements.in
packaging==25.0
    # via
    #   lazy-loader
    #   matplotlib
    #   scikit-image
pillow==12.0.0
    # via
    #   imageio
    #   matplotlib
    #   scikit-image
    #   torchvision
pyparsing==3.2.5
    # via matplotlib
python-dateutil==2.9.0.post0
    # via matplotlib
scikit-image==0.25.2
    # via -r requirements.in
scipy==1.16.3
    # via scikit-image
six==1.17.0
    # via python-dateutil
sympy==1.14.0
    # via torch
tifffile==2025.10.16
    # via scikit-image
torch==2.9.0
    # via
    #   -r requirements.in
    #   torchvision
torchvision==0.24.0
    # via -r requirements.in
tqdm==4.67.1
    # via -r requirements.in
typing-extensions==4.15.0
    # via torch

# The following packages are considered to be unsafe in a requirements file:
# setuptools
